Changed path structure and used json instead of .txt
I changed the path/folder structure, so that everything that belongs directly to the code is in `resources\` and the code in `resources\src\`, whose variables get declared with items form `resources\data\` and under-folders.

Also, I started using json files instead of txt files, which make the code much easier and shorter, but do not work right now.

Co-Authored-By: Josey Wörmann <67111271+joseywoermann@users.noreply.github.com>